+++ test/unit-test.js
@@ -5,10 +5,6 @@ var AssetsInclude = require('../index');
 
 var includeContext = function(definition, options) {
   definition.topic = function() {
-    options = options || {};
-    if (!options.root)
-      options.root = path.join('test', 'data', 'public');
-
     return new AssetsInclude(path.join('test', 'data', 'config.yml'), options);
   };
   return definition;